+++ jax/interpreters/xla.py
@@ -1293,11 +1293,3 @@ pe.staged_out_calls.add(xla_call_p)
 def omnistaging_enabler() -> None:
   global _pval_to_result_handler
   del _pval_to_result_handler
-
-  def _axis_index_translation_rule(c, *, axis_name, axis_env, platform):
-    div = xb.constant(c, np.array(axis_env.nreps // prod(axis_env.sizes),
-                                  dtype=np.uint32))
-    mod = xb.constant(c, np.array(axis_env.sizes[-1], dtype=np.uint32))
-    unsigned_index = xops.Rem(xops.Div(xops.ReplicaId(c), div), mod)
-    return xops.ConvertElementType(unsigned_index, xb.dtype_to_etype(np.int32))
-  parallel_translations[core.axis_index_p] = _axis_index_translation_rule  # type: ignore
